requirements.txt
pandas>=1.3.0
numpy>=1.20.0
yfinance>=0.2.4
ta>=0.10.1
pytz>=2022.1
schedule>=1.1.0
requests>=2.27.0
pykrx>=1.0.39
PyJWT>=2.3.0
python-telegram-bot>=13.12
openai>=1.0.0
python-dotenv>=0.20.0
pytest>=7.0.0